#COBUMMFO-81# Коментар: Реліз 39.3.1

+++ web_rozdrib/ru/barsroot/App_Code/Areas/InsUi/Models/Transport/PaymentsResponse.cs
@@ -15,6 +15,6 @@ namespace BarsWeb.Areas.InsUi.Models.Transport
         /// <summary>
         /// Номер документа(при успішному виконанні)
         /// </summary>
-        public decimal? externalId { get; set; }
+        public Int64? externalId { get; set; }
     }
 }
